move highlight plugin into its own file
